Move message DAO tests to MessagesDAOTest.java
